Unit tests for checking to make sure we can filter HK data on app source  (#124)
* Unit tests for checking to make sure we can filter HK data on app source

Moved the unit tests to LoopKit where they can use `LoopKit Example` (which has HK entitlements)
(Thanks again to Darin for pointing this out)

* Create new LoopKitHostedTests target for running HK tests

The reason this was needed was because the HKFiltering tests need to run in a "host" app that has the HealthKit entitlement.  LoopKit Example has that.  However, we couldn't just make all of LoopKitTests use that hosted app because it causes other problems.

(once again, Darin's idea...thanks, Darin!)

* PR Feedback

* More PR Feedback