LOOP-1406: Add authentication when saving therapy settings (#195)
* LOOP-1406: Add authentication when saving therapy settings

This refactors a bit of the "saving" logic to have common code for presenting the authentication challenge, and then presenting an alert if the challenge fails.

* Don't auth challenge in acceptance flow

* PR Feedback

* Oops.  PR feedback.

* Remove SettingSaveAlert

Now that we don't want to show an alert due to auth failure, it is unnecessary.

* Get rid of extraneous print statements

* remove print statements

* Adds "cancel" button to all of the therapy settings editors

Per the slack thread, it was determined that implying that "back" would cancel the user's edited setting, using "Cancel" was determined to be more clear.  So I've added a cancel button to each editor, but it meant also wiring a cancel() function in so that the parent NavigationLink could navigate back.

* Another tweak to the "cancel" behavior

Another "implicit" design requirement was that the "Cancel" button only show if the value is edited, but that the "back" button should still show if it isn't.  This changes that in all screens.

* PR Feedback

* Move the cancel closure to an environment

Thanks for the suggesiton, @mpangburn!  A lot cleaner...

* make the whole card tappable