Unify language around fast/rapid acting insulins
Fast and rapid-acting appear to be interchangeable in diabetes literature, but the descriptions for Insulin Type and Insulin Model should be consistent when referencing them. I chose to go with Rapid-acting because that is the terminology I see on the ADA website (https://www.diabetes.org/healthy-living/medication-treatments/insulin-other-injectables/insulin-basics). Some manufacturers used rapid, while others use fast.