Restore observer queries startup in HealthKit Stores.
This restores calls to createQuery() lost in the 8947c0f merge.  At
present, these exist in `master` within HealthKitSampleStore's init, but
were moved into individual stores as part of 2ea6be07, and then removed
later for an unknown reason.

This adds tests for all stores that an observer query has been started
during initialization in the case that authorizationRequired returns
false; if authorization is required, queries will be created as part of
authorize().

Fixes #1593